move extensions into parent folder
