+++ FluentValidation/Web/Validators/WeatherForecastValidator.cs
@@ -1,7 +1,6 @@
 ﻿using FluentValidation;
 using Woody230.FluentValidation.Validators;
-using Woody230.FluentValidation.Validators.Extensions;
-using Woody230.FluentValidation.Validators.Property.Extensions;
+using Woody230.FluentValidation.Validators.Property;
 using Woody230.FluentValidation.Web.Models;
 
 namespace Woody230.FluentValidation.Web.Validators;
